Angular_2/prueba/src/app/servicios/alumno.service.ts
import { Injectable } from '@angular/core';
import { Observable, throwError, from } from 'rxjs';
import { HttpClient, HttpErrorResponse } from '@angular/common/http';
import { HttpHeaders } from '@angular/common/http';
import {Alumno} from '../modelos/Alumno';
import { headersToString } from 'selenium-webdriver/http';
@Injectable({
  providedIn: 'root'
})
export class AlumnoService {

  constructor(private http:HttpClient) {}
  addAlumno(alumno:Alumno):Observable<Alumno>{
    return this.http.post<Alumno>('http://homestead.test/api/auth/registroAlumno',alumno,this.getoken());
  }

  getoken(){

    var httpOptions = {
    headers: new HttpHeaders({
      'Content-Type':  'application/json',
      'Authorization': 'Bearer '+localStorage.getItem('accto')
      })
    };
    return httpOptions;
  }

}
